Passing error and route tests, scaffolding for socket tests
