+++ __tests__/gameplay.test.js
@@ -1,0 +1,41 @@
+'use strict';
+
+let client = require('socket.io-client');
+const gameClient = require('../lib/game/game-client.js');
+const gameScreen = require('../lib/game/gameplay-screen.js');
+
+let socket = client.connect();
+
+
+it.skip('should emit getCategoryQuestions event at right time', () => {
+
+});
+
+
+it.skip('should emit doneGettingCats event at right time', () => {
+
+});
+
+
+it.skip('should emit nextQuestion event at right time', () => {  
+
+});
+
+
+it.skip('should emit wrongAnsEvent event at right time', () => {
+
+});
+
+
+
+
+
+
+
+
+
+it('should emit afterEndRender event at right time', () => {
+
+  
+
+});
