+++ __tests__/start-game.test.js
@@ -1,0 +1,17 @@
+'use strict';
+
+let client = require('socket.io-client');
+const gameStart = require('../lib/game/start.js');
+let socket = client.connect();
+
+
+it('should emit usernamePopulate event at right time', () => {
+
+  
+
+});
+
+
+
+
+
